move wrap creation inside render and restructure tests
