+++ packages/rum-react/test/e2e/data.json
