--- packages/rum-react/test/e2e/react/index.html
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-  <title>React test</title>
-  <base href="/test/e2e/react/">
-  <!-- FIXME REMOVE if it is not needed
-  <meta http-equiv="Content-Security-Policy" content="default-src 'self';img-src 'self' data:; script-src 'self' 'unsafe-inline' 'unsafe-eval'; connect-src http://localhost:8000 http://localhost:8200 http://localhost:8001 http://apm-server:8200">
-  -->
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <link rel="icon" href="data:;base64,iVBORw0KGgo=" />
-</head>
-
-<body>
-    <div id="app"></div>
-</body>
-<script src="/test/e2e/react/app.e2e-bundle.js"></script>
-
-</html>
