+++ packages/rum-react/test/e2e/with-router/index.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <title>React test</title>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <link rel="icon" href="data:;base64,iVBORw0KGgo=" />
+  </head>
+
+  <body>
+    <div id="app"></div>
+  </body>
+  <script src="/test/e2e/with-router/general.e2e-bundle.js"></script>
+</html>
